+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/codec/LuceneOptimizedNormsFormatTest.java
@@ -27,6 +27,7 @@ import org.apache.lucene.index.BaseIndexFileFormatTestCaseUtils;
 import org.apache.lucene.index.BaseNormsFormatTestCase;
 import org.apache.lucene.util.TestRuleLimitSysouts;
 import org.junit.BeforeClass;
+import org.junit.Ignore;
 
 import java.io.IOException;
 
@@ -73,5 +74,12 @@ public class LuceneOptimizedNormsFormatTest extends BaseNormsFormatTestCase {
     @Override
     public void testMultiClose() throws IOException {
         BaseIndexFileFormatTestCaseUtils.testMultiClose(this);
+    }
+
+    @Override
+    @Ignore // Issue #2598: Make Lucene @Nightly tests pass for fixed seed
+    @Nightly
+    public void testRamBytesUsed() throws IOException {
+        super.testRamBytesUsed();
     }
 }